Move pkg/listeners away
It is only used in `daemon` and should really live there.
