+++ daemon/listeners/listeners_solaris.go
